// server/src/modules/chat/services/twitch-channel-event-type.enum.ts
export enum TwitchChannelEventType {
    NewFollower = 'newFollower',
    ChannelPointRedeem = 'channelPointRedeem',
    Raid = 'raid',
    Cheer = 'cheer',
    Subscription = 'subscription',
    StreamOnline = 'streamOnline',
    HypeTrainBegin = 'hypeTrainBegin',
    HypeTrainProgress = 'hypeTrainProgress',
    HypeTrainEnd = 'hypeTrainEnd',
    PollBegin = 'pollBegin',
    PollProgress = 'pollProgress',
    PollEnd = 'pollEnd',
}